+++ designer/client/src/containers/theme/nuTheme.tsx
@@ -1,5 +1,5 @@
 import { alpha, createTheme, Palette, PaletteMode } from "@mui/material";
-import { fontFamily, globalStyles } from "./styles";
+import { fontFamily, formLabelWidth, globalStyles } from "./styles";
 import { blendDarken, blendLighten } from "./helpers";
 import { deepmerge } from "@mui/utils";
 import { lightModePalette } from "./lightModePalette";
@@ -190,7 +190,7 @@ export const nuTheme = (mode: PaletteMode, setMode: Dispatch<SetStateAction<Pale
                             ...theme.typography.body2,
                             display: "flex",
                             marginTop: "9px",
-                            flexBasis: "20%",
+                            flexBasis: formLabelWidth,
                             maxWidth: "20em",
                             overflowWrap: "anywhere",
                         }),